requirements.txt
langchain
langchain-openai
langchain-community
tiktoken
ipdb
python-dotenv
spotipy